feat: isolate useFileAsDataURL hook for file preview in modal and composer
