+++ apps/meteor/client/views/room/modals/FileUploadModal/MediaPreview.tsx
@@ -1,35 +1,13 @@
 import { AudioPlayer, Box, Icon } from '@rocket.chat/fuselage';
 import type { ReactElement } from 'react';
-import { useEffect, useState, memo } from 'react';
+import { memo } from 'react';
 import { useTranslation } from 'react-i18next';
 
 import { FilePreviewType } from './FilePreview';
 import ImagePreview from './ImagePreview';
 import PreviewSkeleton from './PreviewSkeleton';
 import { userAgentMIMETypeFallback } from '../../../../lib/utils/userAgentMIMETypeFallback';
-
-type ReaderOnloadCallback = (url: FileReader['result']) => void;
-
-const readFileAsDataURL = (file: File, callback: ReaderOnloadCallback): void => {
-	const reader = new FileReader();
-	reader.onload = (e): void => callback(e?.target?.result || null);
-
-	return reader.readAsDataURL(file);
-};
-
-const useFileAsDataURL = (file: File): [loaded: boolean, url: null | FileReader['result']] => {
-	const [loaded, setLoaded] = useState(false);
-	const [url, setUrl] = useState<FileReader['result']>(null);
-
-	useEffect(() => {
-		setLoaded(false);
-		readFileAsDataURL(file, (url) => {
-			setUrl(url);
-			setLoaded(true);
-		});
-	}, [file]);
-	return [loaded, url];
-};
+import { useFileAsDataURL } from '../../hooks/useFileAsDataURL';
 
 type MediaPreviewProps = {
 	file: File;
